Merge origin/feat/structured-compression-columns to pick up a
follow-up commit pushed after this branch's conflict resolution began:
bigint(20) unsigned for the byte columns, excluding negative deltas
from getTotalBytesSaved(), and no longer fabricating a compression
date for legacy rows migrated by the upgrade wizard.

+++ ext_tables.sql
@@ -1,26 +1,26 @@
 CREATE TABLE sys_file
 (
-    compressed              tinyint(1)   DEFAULT '0' NOT NULL,
-    compress_skipped        tinyint(1)   DEFAULT '0' NOT NULL,
+    compressed              tinyint(1)          DEFAULT '0' NOT NULL,
+    compress_skipped        tinyint(1)          DEFAULT '0' NOT NULL,
     compress_error          text,
-    compress_info           varchar(255) DEFAULT '' NOT NULL,
-    backup_path             varchar(255) DEFAULT '' NOT NULL,
-    compress_provider       varchar(32)  DEFAULT '' NOT NULL,
-    compress_tool           varchar(32)  DEFAULT '' NOT NULL,
-    compress_original_size  int(11)      DEFAULT '0' NOT NULL,
-    compress_size           int(11)      DEFAULT '0' NOT NULL,
-    compress_tstamp         int(11)      DEFAULT '0' NOT NULL,
+    compress_info           varchar(255)        DEFAULT '' NOT NULL,
+    backup_path             varchar(255)        DEFAULT '' NOT NULL,
+    compress_provider       varchar(32)         DEFAULT '' NOT NULL,
+    compress_tool           varchar(32)         DEFAULT '' NOT NULL,
+    compress_original_size  bigint(20) unsigned DEFAULT '0' NOT NULL,
+    compress_size           bigint(20) unsigned DEFAULT '0' NOT NULL,
+    compress_tstamp         int(11)             DEFAULT '0' NOT NULL,
     KEY idx_compressed (compressed)
 );
 
 CREATE TABLE sys_file_processedfile
 (
-    compressed              tinyint(1)   DEFAULT '0' NOT NULL,
+    compressed              tinyint(1)          DEFAULT '0' NOT NULL,
     compress_error          text,
-    compress_provider       varchar(32)  DEFAULT '' NOT NULL,
-    compress_tool           varchar(32)  DEFAULT '' NOT NULL,
-    compress_original_size  int(11)      DEFAULT '0' NOT NULL,
-    compress_size           int(11)      DEFAULT '0' NOT NULL,
-    compress_tstamp         int(11)      DEFAULT '0' NOT NULL,
+    compress_provider       varchar(32)         DEFAULT '' NOT NULL,
+    compress_tool           varchar(32)         DEFAULT '' NOT NULL,
+    compress_original_size  bigint(20) unsigned DEFAULT '0' NOT NULL,
+    compress_size           bigint(20) unsigned DEFAULT '0' NOT NULL,
+    compress_tstamp         int(11)             DEFAULT '0' NOT NULL,
     KEY idx_compressed (compressed)
 );
